Binary kernel modules without kernel versioning
Refresh the binary kernel modules. They were built last time with
kernel which was configured to use auto-versioning, so they had to be
rebuilt all the time to match it. Now that we disabled that option in
kernel, rebuild the modules too one last time and from now on they
should always match.