Update the codec binaries to use /opt/ticel as fixed path
Before this change the codec binaries used ./ as path, so we were
forced to run vlc or any program that used them from /opt/ticel (where
they are installed).
Since we install them into that fixed location, we may as well put
that into the codecs themselves and allow the apps to be run from
anywhere, which is a lot more comfortable.