Allow codecs to load also from yaffs2 (flash) boot.
Up until now there was a bug preventing the TI codecs to be able to
load on flash-based boots (using yaffs2 file system).
It worked only on NFS netboot.
This patch fix dsplink so that it is able to load the codec files also
when booting from yaffs, by correcting an error in the way it used
kernel-space file access functions.